Make admin and bootstrap-osd key creation idempotent and consistent.
Avoids a problem with multi-monitor setups where each ceph-mon node
creates a different key.

As a side effect, the key creation is deferred until after ceph-mon
is in quorum. This means it may need another chef-client run to happen.
That's unfortunate but currently unavoidable. The client-admin key
creation could be done outside of chef, but the bootstrap-osd key can
only be stored in chef attributes at the time of a chef-client run.