don't force node attributes to have been set for the Ceph cluster
Instead, pass them in to get_mon_nodes. Push up the previous node
attributes to the mon and osd recipes, but let the client recipe
run searches based on whether we're in crowbar or not.

The crowbar branch is a hack that needs to be replaced with a
proper search of the available choices, instead of assuming it's
for nova...