Pool LWRP and auto create user-defined pools
This change introduces a pool LWRP that allows Ceph user pools to be
created and/or deleted by simply specifying their name and number of
placement groups.

It also allows users to define their own pools in Chef attributes or
environments that are then created automatically when nodes are
provisioned.