Fix timing issue with ceph pool create
There are scenarios being encountered where ceph::mon attempts to create
pools before all the required ceph services are up and running. This
causes the recipe to fail.

My proposed solution to this issue is to remove the 'ceph pool create'
block from the ceph::mon recipe and place it in it's own separate recipe
(ceph::pools), which is placed below ceph::mon in the runlist for the role
(ceph-mon.json).

Having this code in a separate recipe allows better control of when the
pools are created.