Add bootstrap-osd key after it's been created if databags are not used
This change fixes a regression introduced 807f08 that caused
the cookbook to fail if data_bags were not used. The bootstrap-osd
key would not be created at the time of the key being added to the
keyring, causing the chef run to fail. We now try to add the key after
we've checked if the key needs to be created, which will only happen
if we're not using data_bags